create useDynamicColor hook
