+++ src/lib/components/dynamicBackground/DynamicBackground.tsx
@@ -1,10 +1,6 @@
-import React, { FC, useLayoutEffect, useRef } from "react";
-import useDirection from "../../hooks/useDirection";
-import useProximity from "../../hooks/useProximity";
-import { Directions } from "../../types";
-import getSteps from "../../utils/calculations/getSteps";
-import clamp from "../../utils/calculations/clamp";
+import React, { FC, useRef } from "react";
 import { BackgroundProps } from "./types";
+import useDynamicColor from "../../hooks/useDynamicColor";
 
 const DynamicBackground: FC<BackgroundProps> = ({
   startColor = [0, 0, 0],
@@ -13,53 +9,23 @@ const DynamicBackground: FC<BackgroundProps> = ({
   style = {},
   className = "",
 }) => {
-  const ref = useRef(null);
-  const steps = useRef<number[]>([1, 1, 1]);
-  const color = useRef<number[]>([...startColor]);
+  const elementRef = useRef(null);
 
-  const { onSight, y } = useProximity(ref);
-  const direction = useDirection();
-
-  useLayoutEffect(() => {
-    steps.current = getSteps(startColor, endColor, ref.current!);
-  }, []);
-
-  useLayoutEffect(() => {
-    switch (direction) {
-      case Directions.up: {
-        if (onSight && color.current.some((color, i) => color != startColor[i]))
-          color.current = color.current.map((color, i) =>
-            clamp(startColor[i], endColor[i], color - steps.current[i])
-          );
-        break;
-      }
-      case Directions.down: {
-        if (onSight && color.current.some((color, i) => color != endColor[i]))
-          color.current = color.current.map((color, i) =>
-            clamp(startColor[i], endColor[i], color + steps.current[i])
-          );
-        break;
-      }
-    }
-  }, [y]);
+  const color = useDynamicColor({ startColor, endColor, elementRef });
 
   return (
     <div
-      ref={ref}
+      ref={elementRef}
       style={{
         height: "inherit",
         width: "100%",
         ...style,
-        backgroundColor: `rgb(${color.current[0]}, ${color.current[1]}, ${color.current[2]})`,
+        backgroundColor: color,
       }}
       className={className}
       data-testid="dynamic-background"
     >
-      {typeof children == "function"
-        ? children(
-            `rgb(${color.current[0]}, ${color.current[1]}, ${color.current[2]})`
-          )
-        : children}
+      {typeof children == "function" ? children(color) : children}
     </div>
   );
 };
